+++ go-test/constants.go
@@ -1,0 +1,16 @@
+package main
+
+const (
+	EndpointCreateTemplate       = "/api/template/create"
+	EndpointDeleteImage         = "/api/image/delete"
+	EndpointNodeInfo            = "/api/node/info"
+	EndpointNodeAdd             = "/api/node/add"
+	EndpointLVMCreate           = "/api/storage/lvm/create"
+	EndpointNFSCreate           = "/api/storage/nfs/create"
+	EndpointNFSUpdate           = "/api/storage/nfs/update"
+	EndpointCreateVM            = "/api/vm/create"
+	EndpointAttachSecurityGroup = "/api/security-group/attach"
+	EndpointDetachSecurityGroup = "/api/security-group/detach"
+	EndpointBackupCreate        = "/api/backup/create"
+	EndpointBackupRestore       = "/api/backup/restore"
+)
