+++ go-test/main.go
@@ -32,44 +32,3 @@ func main() {
 	log.Fatal(http.ListenAndServe(handlers.Port, nil))
 }
 
-func registerHandlers() {
-	// ===== テンプレート & イメージ =====
-	http.HandleFunc(EndpointCreateTemplate, handlers.HandleCreateTemplate)
-	http.HandleFunc(EndpointDeleteImage, handlers.HandleDeleteImage)
-
-	// ===== ノード =====
-	http.HandleFunc(EndpointNodeInfo, handlers.HandleNodeInfo)
-	http.HandleFunc(EndpointNodeAdd, handlers.HandleNodeAdd)
-
-	// ===== ストレージ - LVM =====
-	http.HandleFunc(EndpointLVMCreate, handlers.HandleLVMCreate)
-
-	// ===== ストレージ - NFS =====
-	http.HandleFunc(EndpointNFSCreate, handlers.HandleNFSCreate)
-	http.HandleFunc(EndpointNFSUpdate, handlers.HandleNFSUpdate)
-	// ===== 仮想マシン =====
-	http.HandleFunc(EndpointCreateVM, handlers.HandleCreateVM)
-
-	// ===== ファイアウォール / セキュリティグループ =====
-	http.HandleFunc(EndpointAttachSecurityGroup, handlers.HandleSecurityGroupAttach)
-	http.HandleFunc(EndpointDetachSecurityGroup, handlers.HandleSecurityGroupDetach)
-
-	// ===== バックアップ =====
-	http.HandleFunc(EndpointBackupCreate, handlers.HandleBackupCreate)
-	http.HandleFunc(EndpointBackupRestore, handlers.HandleBackupRestore)
-}
-
-const (
-	EndpointCreateTemplate       = "/api/template/create"
-	EndpointDeleteImage         = "/api/image/delete"
-	EndpointNodeInfo            = "/api/node/info"
-	EndpointNodeAdd             = "/api/node/add"
-	EndpointLVMCreate           = "/api/storage/lvm/create"
-	EndpointNFSCreate           = "/api/storage/nfs/create"
-	EndpointNFSUpdate           = "/api/storage/nfs/update"
-	EndpointCreateVM            = "/api/vm/create"
-	EndpointAttachSecurityGroup = "/api/security-group/attach"
-	EndpointDetachSecurityGroup = "/api/security-group/detach"
-	EndpointBackupCreate        = "/api/backup/create"
-	EndpointBackupRestore       = "/api/backup/restore"
-)
